Crate netfw

Re-exports

pub use self::policy::FirewallPolicy;

pub use self::rule::FirewallRule;

pub use self::rules::FirewallRules;

Modules

policy

rule

rules

Structs

FirewallProfile

FirewallRulesIter

VariantEnumerator

Enums

FirewallAction

FirewallRuleDirection

Functions

bstr_to_os_string⚠
Panics if bstr is null or bstr data length in bytes is not a multiple of 2

os_str_to_bstr