Function netfw::bstr_to_os_string

pub unsafe fn bstr_to_os_string(bstr: BSTR) -> OsString

Panics if bstr is null or bstr data length in bytes is not a multiple of 2

Safety

bstr must be a valid BSTR.